Module Ostap

module Combinators : sig ... end

module Errors : sig ... end

module Matcher : sig ... end

module Msg : sig ... end

Parsing message interface.

module Pretty : sig ... end

Pretty --- a general set of pretty-printer combinators. Implemented as a wrapper for the standard module Format to ease its use.

module PrioReorderer : sig ... end

Reordering expression in according to priorities of its operators.

module Reason : sig ... end

Predefined implementation of reason type to describe various errors happened during parsing.

module Trap : sig ... end

Trap is infrastructure to maintain the relation between parsed items and their locator information.

module Types_ : sig ... end

module Util : sig ... end