Module Msg.Locator

Various ways to denote the location in the source text.

type t =

| No
(*
No location defined
*)

| Point of Coord.t
(*
One point in the text
*)

| Interval of Coord.t * Coord.t
(*
Contiguous interval of points
*)

| Set of t list
(*
Non-contiguous set of locations
*)

Locator type.

val makeInterval : t -> t -> t

Makes simple interval of two points or set of two non-point locators.

val least : t -> Coord.t

val most : t -> Coord.t

val updateToString : FileLoc.r -> string -> unit

val toString : t -> string

String conversion.

val compare : t -> t -> int

Comparison function.