Module type PrioReorderer.Expression

Signature to abstract expression being reordered.

type t

The type of expression.

val discover : t -> [ `Infix of int * t * t | `Other ]

Discovers the kind of expression: either infix with priority and two arguments or other kind of expression.

val replace : t -> t -> t -> t

replace expr x y replaces arguments of infix expression expr with x and y

val map : (t -> t) -> t -> t

Map function for expressions.