Module Pa_ostap.Extension

Pa_ostap --- a camlp5 syntax extension for BNF-like grammar definitions.

General description

Pa_ostap extends Objective Caml grammar with construct ostap (..), introduced at structure and expression levels. The exact allowed forms are the following:

ostap doc_tag (rules) --- at the structure (module implementation) level;

let ostap doc_tag (rule) --- at the let-binding level; introduces recursive binding;

ostap doc_tag (rules) --- at the expression level;

ostap (parse_expr) --- at the expression level;

let ostap doc_tag (rules) in expr --- at the expression level;

In the specification above rules denotes a sequence of grammar rules, rule --- a single rule, parse_expr --- parse expression, doc_tag --- documentation specifier. Below the examples of all these constructs are given:

(* Grammar rules specification at the structure level; rules are mutually recursive *)
ostap (
  x: IDENT; (* rule defining parser x *)
  y: CONST  (* rule defining parser y *)
)

(* Grammar rule at the let-binding level; bindings are mutually recursive *)
let ostap (x: IDENT) (* rule defining parser x *)
and ostap (y: CONST) (* rule defining parser y *)
and u = 3            (* an example to demonstrate interoperability with other let-bindings *)

let _ =
  (* Let-bindings at expression level; x and y are mutually recursive *)
  let ostap (x: IDENT; y: CONST) in
  (* Grammar expression *)
  let p = ostap (x y) in
  ()

All these constructs are converted into pure OCaml using Ostap parser combinators.

While Ostap is purely abstract with regard to stream implementation Pa_ostap additionally provides convenient integration of parsing and lexing by considering streams as objects. Namely, the stream of tokens L1, L2, ..., Lk is represented by an object with member functions getL1, getL2, ..., getLk. Such a representation allows to freely combine various parser functions that operate on different types of streams with almost no type limitations at construction time.

Additionally to this documentation we provide a closed example of how to use Pa_ostap (see sample directory of the distribution).

Parse expressions

The syntax of parse expressions is as follows (text in bold denotes meta-language syntax description symbols):

parse_expr : alternative1 | alternative2 | ... | alternativek

alternative : prefixed+ [ semantic ]

prefixed : [ - ] basic

basic : [ binding ] postfix [ predicate ]

postfix : primary | postfix ( * [ folding ] | + [ folding ] | ? | :: ( EXPR ) )

folding : with { EXPR } { EXPR }

primary : UIDENT | parser [ parameters ] | string | $ | ( parse_expr )

parser : LIDENT | !(EXPR)

string : STRING | $(EXPR)

parameters : ([ EXPR ])*

binding : PATT :

predicate : => { EXPR } [ ::( EXPR )] =>

semantic : { EXPR }

Here UIDENT and LIDENT stand for identifiers starting from uppercase and lowercase letters correspondingly, STRING --- for OCaml strings, EXPR --- for OCaml expression, PATT --- for OCaml pattern.

parser within parse expression denotes a parse function that applied to a stream to obtain parsed value and residual stream (see module Ostap.Combinators). Each reference is either a LIDENT or arbitrary OCaml expression of appropriate type, surrounded by !(...). Parser invocation may be equipped with parameters each of which has to be surrounded by [...] (partial application is allowed as well). UIDENT is treated as a lexeme reference; thought generally speaking parsing with Ostap does not require any lexer to be provided (you may instead supply a set of basic parse functions in any way you find convenient) Pa_ostap additionally operates with some predefined representation of streams as objects (see module Matcher). This representation does not interfere with the common approach and you need not use this feature unless you explicitly apply to it. There are only three constructs that refer to object implementation of streams: UIDENT, $(EXPR) and STRING. If you use UIDENT in grammar expression, for example NAME, then the stream to parse with this expression has to provide a member function getNAME. Similarly using STRING in expression requires stream to provide a member look. Finally you may match a stream against value of any OCaml expression of string type by surrounding it with $(...).

Postfix operators +, * and ? denote respectively one-or-more iteration, zero-or-more iteration and optional value. Postfix operator ::(EXPR) can be used to comment the reason returned on failure with the given reason value (see Ostap.Combinators.comment function and module Reason as reference implementation).

Additionally some folding can be specified for postfix + and * operators. The folding has the form with { EXPR}{ EXPR } where the first expression in curved brackets denotes initial value for folding with function given by the second expression. For example

callee:expression call:(-"(" arguments -")")* with{callee}{fun callee args -> `Call (callee, args)} {call}

is equivalent to

callee:expression call:(-"(" arguments -")")* {List.fold_left (fun callee args -> `Call (callee, args)) callee call}

Symbol $ within parse expression serves as a shortcut for Ostap.Combinators.lift and so delivers underlying stream as its semantic value.

Prefix operator - is used to omit parsed value from the result of parsing (the parsing of its operand however is not omitted).

Prefix construct PATT: is used to match successfully parsed value against pattern PATT; this matching may provide bindings that can be used later.

Construct =>{EXPR}=> is used to supply additional check of successfully parsed value; EXPR has to be of boolean type and may use bindings made before.

We will not describe the meaning of all constructs in all details since generally it follows the common BNF style; instead we demonstrate some examples that cover all cases of their exploration.

Examples:

"(" expression ")" is a grammar expression to define a function that matches a stream against successive occurrences of "(", that that parsed by expression, and ")". On success this function returns a triple: the token for "(" (of type determined by stream implementation), the value parsed by expression, and the token for ")". There are generally two ways to exclude "(" and ")" from the result. The first way is to bind the result of expression to some name and then explicitly specify the result of grammar expression as follows:

"(" e:expression ")" {e}

The second is just to say to omit brackets:

-"(" expression -")".

Note that you may specify arbitrary pattern in the left part of binding. Prefix omitting operator "-" may also be applied to any grammar expression, enclosed in brackets.

hd:item tl:(-"," item)* {hd :: tl} defines a function to parse a list of items.

(s:string {`Str s} | x:integer {`Int x})* defines a function to parse a list of strings or integers.

hd:integer tl:(-(","?) integer)* {hd :: tl} parses a list of integers delimited by optional commas.

x:integer => {x > 0}::("positive value expected") => {x} parses positive integer value.

x:(integer?) => {match x with Some 0 -> false | _ -> true} => {x} parses optional non-zero integer value.

x:!(MyParseLibrary.MyModule.parseIt) parses a stream with parse function specified by qualified name.

In all examples above we assume that integer parses integer value, string --- string value.

Rules

Rule is named and optionally parameterized parse expression; several mutually-recursive rules may be defined at once. The syntax of rule definition is

rule : LIDENT arguments : [ predicate ] parse_expr

rules : rule1; rule2; ...; rulek

arguments : ([PATT] )*

For example,

ostap (
  sequence[start]: item[start] | next:item[start] sequence[next];
  item[start]: x:integer {x+start} | ";" {start};
  entry: sequence[0]
)

declares (among others) the parser function entry which parses and sums a semicolon-terminated sequence of integers.

Documentation generation

Option -tex filename makes Pa_ostap generate LaTeX documentation for all rules. On default all text is placed into specified file; however the output can be split into several files by specifying doc_tag option for ostap construct. The syntax of option is as follows:

doc_tag : [ [ STRING ] ]

With this option provided the documentation for corresponding rules will be placed in file with name filename.tagname.tex, where filename is the name specified by option -tex and tagname is string value of doc_tag.

Generated documentation uses ostap.sty package which provides cross-references and automatic layout for most of the cases. To obtain good documentation it is recommended to use simple parsers (i.e. identifiers) in grammar rules. Parameterized rules are supported as well.