Package fix 20230505

fix is an OCaml library that helps with various algorithmic constructions that involve memoization, recursion, and numbering.

Installation and Usage

Type opam install fix.

In your dune file, add (libraries fix) to the description of your library or executable.

Within your code, you may wish to declare open Fix. This allows you to access the submodules without the leading "Fix." qualifier.

Data Flow Analysis

The following submodules help solve systems of recursive monotone equations. In other words, they help implement data flow analyses.

Fix.Fix

Fix.DataFlow

The following submodules help construct the arguments required by the functors in Fix.Fix and Fix.DataFlow.

Fix.Prop

Fix.Glue

Numbering

The following submodules help generate unique numbers and assign unique numbers to objects.

Fix.Gensym

Fix.HashCons

Fix.Numbering

Fix.GraphNumbering

Fix.Indexing

Memoization and Tabulation

The following submodules help construct memoized or tabulated functions, both of which have the property that repeated computation is avoided.

Fix.Memoize

Fix.Tabulate

Minimization

The following submodules offer minimization algorithms.

Fix.Minimize

Data Structures

The following submodules offer data structures that can be of general interest.

Fix.CompactQueue

Fix.Enum

Fix.Partition

Package info

authors

François Pottier <francois.pottier@inria.fr>

changes-files

CHANGES.md

depends

dune

ocaml

homepage

https://gitlab.inria.fr/fpottier/fix

issues

francois.pottier@inria.fr

license

LGPL-2.0-only

license-files

LICENSE

maintainers

francois.pottier@inria.fr

readme-files

README.md

repo

git+https://gitlab.inria.fr/fpottier/fix.git

version

20230505