Package ocamlformat 0.26.2

If you are here, you are probably interested in using a formatting tool for your code base, so that you do not have to worry about formatting it by hand, and to speed up code review by focusing on the important parts.

OCamlFormat is probably what you are after! OCamlFormat is a tool to format OCaml code.

Table of Content

Getting started

Editor setup

Manpage

How-To's

Frequently asked questions

API

Doc-comments language reference

RPC

Package info

authors

Etienne Millon <etienne@tarides.com>

Guillaume Petiot <guillaume@tarides.com>

Hugo Heuzard <hugo.heuzard@gmail.com>

Josh Berdine <jjb@fb.com>

Jules Aguillon <jules@j3s.fr>

changes-files

CHANGES.md

depends

cmdliner

dune

ocaml

ocamlformat-lib

ocamlformat-rpc-lib

odoc

re

homepage

https://github.com/ocaml-ppx/ocamlformat

issues

https://github.com/ocaml-ppx/ocamlformat/issues

license

LGPL-2.1-only WITH OCaml-LGPL-linking-exception

MIT

license-files

LICENSE

maintainers

Emile Trotignon <emile@tarides.com>

Guillaume Petiot <guillaume@tarides.com>

Jules Aguillon <jules@j3s.fr>

readme-files

README.md

repo

git+https://github.com/ocaml-ppx/ocamlformat.git

version

0.26.2