Package ppxlib 0.33.0

Overview

This is the user manual and API for ppxlib, the core of the PPX meta-programming system for OCaml and its derivatives, such as Reason. For a good introduction on PPXs, what they are, and how to use them, see the OCaml official guide on PPXs. This manual is mostly aimed at authors of PPX rewriters and contains everything one should know in order to write PPX rewriters.

Manual

The manual consists of several sections. It can be read linearly, but you can also jump directly to your section of interest:

An introduction to ppxlib

How ppxlib works internally

Registering a transformation

Generating AST nodes

Destructing AST nodes

Traversing the AST

Good practices

Examples

API

The API exposes the following modules:

Ppxlib

Ppxlib_ast

Astlib

Ppxlib_metaquot

Ppxlib_metaquot_lifters

Ppxlib_print_diff

Ppxlib_runner

Ppxlib_runner_as_ppx

Stdppx

Ppxlib_traverse

Ppxlib_traverse_builtins

Package info

authors

Jane Street Group, LLC <opensource@janestreet.com>

changes-files

CHANGES.md

HISTORY.md

depends

cinaps

dune

ocaml

ocaml-compiler-libs

ocamlfind

odoc

ppx_derivers

re

sexplib0

sexplib0

stdlib-shims

homepage

https://github.com/ocaml-ppx/ppxlib

issues

https://github.com/ocaml-ppx/ppxlib/issues

license

MIT

license-files

LICENSE.md

maintainers

opensource@janestreet.com

online-doc

https://ocaml-ppx.github.io/ppxlib/

readme-files

README.md

repo

git+https://github.com/ocaml-ppx/ppxlib.git

version

0.33.0